tree: make tree interface class
